+++ waterloop-site/src/pages/Sponsors.tsx
@@ -1,15 +1,19 @@
 import React from "react";
-import Hero from "components/Hero/General"
+import Hero from "components/Hero/General";
+import SponsorList from "components/SponsorComponents/SponsorPage";
 
 class Sponsors extends React.Component {
-    render() {
-        return (
-            <div>
-                <Hero title="Support Our Team" description="Help fund the future."></Hero>
-            </div>
-        );
-    }
+  render() {
+    return (
+      <div>
+        <Hero
+          title="Support Our Team"
+          description="Help fund the future."
+        ></Hero>
+        <SponsorList></SponsorList>
+      </div>
+    );
+  }
 }
 
 export default Sponsors;
-
